+++ src/py_logic/user.py
@@ -1,0 +1,6 @@
+class User:
+
+    def __init__(sid):
+        self.sid = sid
+        self.games_playing = []
+        self.games_observing = []